pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "formai"
version = "0.2.0"
description = "Sub-agent based intelligent form processing pipeline for PDF templates and filled forms."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Codex" }
]
dependencies = [
  "Pillow>=10.0.0",
  "PyMuPDF>=1.24.0",
  "pypdf>=5.3.0",
  "reportlab>=4.0.0"
]

[project.optional-dependencies]
vision = [
  "openai>=1.30.0"
]
glm_ocr = [
  "torch>=2.1.0",
  "transformers>=4.57.0",
  "accelerate>=1.0.0"
]
chandra = [
  "chandra-ocr>=0.2.0",
  "transformers>=5.2.0",
  "accelerate>=1.11.0"
]
surya = [
  "surya-ocr>=0.17.0"
]
local_vlm = [
  "sf-hamilton>=1.70.0",
  "chandra-ocr>=0.2.0",
  "surya-ocr>=0.17.0",
  "torch>=2.1.0",
  "transformers>=5.2.0",
  "accelerate>=1.11.0"
]
commonforms = [
  "commonforms @ git+https://github.com/jbarrow/commonforms.git"
]
benchmarks = [
  "datasets>=2.19.0"
]
api = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "python-multipart>=0.0.20"
]
dev = [
  "pytest>=8.0.0",
  "sf-hamilton>=1.70.0"
]

[project.scripts]
formai = "formai.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
